4.2.0

+++ Minecraft_QQ_Core/Robot/RobotCore.cs
@@ -7,7 +7,7 @@ public static class RobotCore
 {
     public static RobotSDK Robot = new();
 
-    private static void Message(byte type, object data)
+    private static void Message(int type, object data)
     {
         switch (type)
         {
